refactor: 핸들러 위치 수정
